MirrorPageRecord: stop relying on the `remote_redirect` table
Rather than determining if a mirrored page is a redirect based on whether it
has a row in the `remote_redirect` table, use a (cached) API call to check.
While this sacrifices a bit of performance, it will allow decoupling mirrored
pages from their page ids on the remote wiki.

SEL-2228