Limit mirroring to pages in `remote_page`
Rather than the "fast" path being to look in that table, and the "slow" path
being to check the remote API, the "slow" path checks both. Thus, the pages
that are mirrored can be limited by controlling what is included in the
`remote_page` table without then causing issues when that table was previously
bypassed in some code paths.

Also add some basic tests for `Mirror::canMirror()`

SEL-2228